ダメージ持続時間をイントに

+++ JapanGameGrandprize_2-4A/JapanGameGrandprize_2-4A/BULLET.h
@@ -6,7 +6,7 @@ class BulletBase : public SphereCollider
 {
 protected:
 	float damage;          //弾のダメージ
-	float debuff_time;	 //弾の効果時間
+	int debuff_time;	 //弾の効果時間
 	int efect_count;	 //エフェクトの時間
 	float scrool_x;      //弾のスクロールの値
 	bool delete_flg;     //弾の削除フラグ
